chore: seperated types and services from custom controls component
